provider_fav_list_and_info_page_placeholder_after_pull
